+++ examples/mdl_sdk/mdl_browser/mdl_qt_plugin/index/index_document.h
@@ -33,6 +33,7 @@
 #ifndef MDL_SDK_EXAMPLES_MDL_BROWSER_DOCUMENT_BASE_H
 #define MDL_SDK_EXAMPLES_MDL_BROWSER_DOCUMENT_BASE_H
 
+#include <cstdint>
 #include <vector>
 #include <string>